+++ R/pd_importance.R
@@ -57,7 +57,7 @@ pd_importance.default <- function(object, ...) {
 #' @describeIn pd_importance PD based feature importance from "hstats" object.
 #' @export
 pd_importance.hstats <- function(object, normalize = TRUE, squared = TRUE, 
-                                 sort = TRUE, top_m = 15L, drop_zero = FALSE, 
+                                 sort = TRUE, top_m = 15L, zero = TRUE, 
                                  eps = 1e-8, plot = FALSE, 
                                  fill = "#2b51a1", ...) {
   num <- with(
@@ -73,7 +73,7 @@ pd_importance.hstats <- function(object, normalize = TRUE, squared = TRUE,
     squared = squared, 
     sort = sort, 
     top_m = top_m,
-    drop_zero = drop_zero,
+    zero = zero,
     eps = eps
   )
   if (plot) plot_stat(out, fill = fill, ...) else out